backend/requirements.txt
fastapi[all]
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
SQLAlchemy
psycopg2